+++ python/CHANGELOG.md
@@ -15,6 +15,7 @@ We structure this changelog in accordance with [Keep a Changelog](https://keepac
 - Add support for ETL context manager.
 - Add `cont_on_err` option for bucket transform.
 - Add `job_id` option to `Etl.view()` method.
+- Add support for OS packages in `init_class`.
 
 ### Changed